requirements.txt
python-dotenv==1.0.*
requests==2.30.*
python-telegram-bot==13.15.*
redis==5.0.*